refactor: Extract Superclass to its own file
